test: move integration tests into crate
