+++ codecov.yml
@@ -7,6 +7,7 @@ coverage:
 
 ignore:
   - "src/bin/**"
+  - "src/integration_tests/**"
   - "tests/**"
   - "benches/**"
   - "examples/**"
