+++ src/integration_tests/common/Dockerfile
