+++ src/integration_tests/mod.rs
@@ -1,0 +1,10 @@
+mod common;
+
+mod clickhouse_test;
+mod delimiter_reader_test;
+mod parquet_output_test;
+mod parser_test;
+mod record_test;
+mod report_test;
+mod source_file_test;
+mod source_standalone_test;
