+++ src/lib.rs
@@ -9,3 +9,6 @@ pub mod parser;
 pub mod record;
 pub mod report;
 pub mod source;
+
+#[cfg(test)]
+mod integration_tests;
