+++ src/parser/runtime/combinators.rs
@@ -1,11 +1,9 @@
 pub mod rdb_str_box;
 pub mod reduce;
 pub mod seq;
-pub mod skip_bytes;
 pub mod varint;
 
 pub use rdb_str_box::RDBStrBox;
 pub use reduce::ReduceParser;
-pub use seq::{ParserPhase, Seq2Parser, Seq3Parser, Seq4Parser, Seq5Parser};
-pub use skip_bytes::SkipBytesParser;
+pub use seq::{Seq2Parser, Seq3Parser, Seq4Parser, Seq5Parser};
 pub use varint::RDBLenParser;
